Types

Types

ModiaMedia.AbstractFluidConstants — Type.

abstract type AbstractFluidConstants - Abstract type of all FluidConstants structures

source

ModiaMedia.AbstractMedium — Type.

abstract type AbstractMedium - Abstract type of all media

source

ModiaMedia.BasicFluidConstants — Type.

fluidConstants = BasicFluidConstants(;iupacName="", casRegistryNumber="", 
                     chemicalFormula="", structureFormula="", molarMass=NaN)

Generate a BasicFluidConstants <: AbstractFluidConstants object containing the minimal information about the standard data of the medium.

source

ModiaMedia.FluidInfos — Type.

infos = FluidInfos(; <keyword arguments, see below>)

Generate a new FluidInfos object, containing generic properties of a medium.

Keyword arguments

Name and type	Description

mediumName::AbstractString	Name of the medium

substanceNames::Vector{AbstractString}	Names of the mixture substances. Set substanceNames=[mediumName] if only one substance.

extraPropertiesNames::Vector{AbstractString}	Names of the additional (extra) transported properties. Set extraPropertiesNames=fill("",0) if unused

ThermoStates::IndependentVariables	Enumeration type for independent variables

baseProperties::Symbol	Symbol of baseProperties model = :BaseProperties_<StructName>

singleState::Bool	= true, if u and d are not a function of pressure

reducedX::Bool	= true if medium contains the equation sum(X) = 1.0; set reducedX=true if only one substance

fixedX::Bool	= true if medium contains the equation X = reference_X

reference_p	Reference pressure of Medium: default 1 atmosphere

reference_T	Reference temperature of Medium: default 25 deg Celsius

reference_X	Reference mass fractions of medium

p_default	Default value for pressure of medium (for initialization)

T_default	Default value for temperature of medium (for initialization)

h_default	Default value for specific enthalpy of medium (for initialization)

X_default	Default value for specific enthalpy of medium (for initialization)

nS	Number of substances

nX	Number of mass fractions

nXi	Number of structurally independent mass fractions

nC	Number of extra (outside of standard mass-balance) transported properties

C_nominal	Default for the nominal values for the extra properties

Example

import ModiaMedia

infos = ModiaMedia.FluidInfos(mediumName           = "simpleMedium",
                              substanceNames       = [mediumName],
                              extraPropertiesNames = fill("",0),
                              ThermoStates         = IndependentVariables_T)

source

ModiaMedia.FluidLimits — Type.

fluidLimits = FluidLimits(; TMIN=NaN, TMAX=NaN, DMIN=NaN, DMAX=NaN, PMIN=NaN, PMAX=NaN, 
                            HMIN=NaN, HMAX=NaN, SMIN=NaN, SMAX=NaN)

Generate a new FluidLimits object, containing the validity limits of the medium.

source

ModiaMedia.IdealGasFluidConstants — Type.

fluidConstants = IdealGasFluidConstants(; iupacName="", casRegistryNumber="", 
      chemicalFormula="", structureFormula="", molarMass=NaN,
      criticalTemperature=NaN, criticalPressure=NaN, criticalMolarVolume=NaN, 
      acentricFactor=NaN, meltingPoint=NaN, normalBoilingPoint=NaN, dipoleMoment=NaN,
      hasIdealGasHeatCapacity=false, hasCriticalData=false, hasDipoleMoment=false,
      hasFundamentalEquation=false, hasLiquidHeatCapacity=false, hasSolidHeatCapacity=false,
      hasAccurateViscosityData=false, hasAccurateConductivityData=false,
      hasVapourPressureCurve=false, hasAcentricFactor=false,
      HCRIT0=0.0, SCRIT0=0.0, deltah=0.0, deltas=0.0)

Generate a IdealGasFluidConstants <: AbstractFluidConstants object containing the minimal information about the standard data of ideal gas media (critical, triple, molecular and other standard data).

source

ModiaMedia.IndependentVariables — Type.

@enum IndependentVariables

Enumeration defining the independent variables of a medium. Possible values:

Value	Independent variables

IndependentVariables_T	Temperature

IndependentVariables_pT	Pressure, temperature

IndependentVariables_ph	Pressure, specific enthalpy

IndependentVariables_phX	Pressure, specific enthalpy, mass fractions

IndependentVariables_pTX	Pressure, temperature, mass fractions

IndependentVariables_dTX	Density, temperature, mass fractions

source

ModiaMedia.Init — Type.

Enumeration Init defining initialization for fluid flow

source

ModiaMedia.PureSubstance — Type.

abstract type PureSubstance <: AbstractMedium - Abstract type of all media consisting of a pure substance

source

ModiaMedia.PureSubstanceThermodynamicState — Type.

abstract type PureSubstanceThermodynamicState <: ThermodynamicState - Abstract type of the states of all media consisting of a pure substance

source

ModiaMedia.ReferenceEnthalpy — Type.

@enum ReferenceEnthalpy

Enumeration defining the reference enthalpy of a medium. Possible values:

ReferenceEnthalpy_ZeroAt0K: The enthalpy is 0 at 0 K (default), if the enthalpy of formation is excluded

ReferenceEnthalpy_ZeroAt25C: The enthalpy is 0 at 25 degC, if the enthalpy of formation is excluded

ReferenceEnthalpy_UserDefined: The user-defined reference enthalpy is used at 293.15 K (25 degC)

source

ModiaMedia.ReferenceEntropy — Type.

@enum ReferenceEntropy

Enumeration defining the reference entropy of a medium. Possible values:

ReferenceEntropy_ZeroAt0K: The entropy is 0 at 0 K (default)

ReferenceEntropy_ZeroAt0C: The entropy is 0 at 0 degC

ReferenceEntropy_UserDefined: The user-defined reference entropy is used at 293.15 K (25 degC)

source

ModiaMedia.SimpleMedium — Type.

medium = SimpleMedium(; mediumName     = Missing,
                        reference_p    = 101325,
                        reference_T    = 298.15,
                        p_default      = 101325,
                        T_default      = 293.15,
                        fluidConstants = nothing, 
                        fluidLimits    = FluidLimits(), 
                        data           = nothing)

Generate a SimpleMedium <: PureSubstance medium object.

source

ModiaMedia.SimpleMediumData — Type.

data = SimpleMediumData(;cp_const=NaN, cv_const=NaN, d_const=NaN, eta_const=NaN,
                         lambda_const=NaN, a_const=NaN, T0=NaN, MM_const=NaN)

Generate a SimpleMediumData object containing the data for a SimpleMedium medium.

source

ModiaMedia.SingleGasNasa — Type.

medium = SingleGasNasa(; mediumName     = Missing,
                         reference_p    = 101325,
                         reference_T    = 298.15,
                         p_default      = 101325,
                         T_default      = 293.15,
                         fluidConstants = nothing, 
                         fluidLimits    = FluidLimits(TMIN=200.0, TMAX=6000.0), 
                         data           = nothing)

Generate a SingleGasNasa <: PureSubstance medium object.

source

ModiaMedia.SingleGasNasaData — Type.

data = SingleGasNasaData(;name=Missing, MM=NaN, Hf=NaN, H0=NaN, Tlimit=NaN,
                          alow=Missing, blow=Missing, ahigh=Missing,
                          bhigh=Missing, R=NaN)

Generate a SingleGasNasaData object containing the data for an ideal Gas based on the NASA Glenn coefficients.

source

ModiaMedia.Th — Type.

Enumeration Th - defining whether T or h are known as boundary condition

source

ModiaMedia.ThermodynamicState — Type.

abstract type ThermodynamicState - Abstract type of all media states

source

ModiaMedia.ThermodynamicState_pT — Type.

state = ThermodynamicState_pT(p,T)

Generate a ThermodynamicState_pT <: PureSubstanceThermodynamicState object containg pressure p [Pa] and temperature T [K] as states.

source

ModiaMedia.pd — Type.

Enumeration pd defining whether p or d are known for the boundary condition

source