Exported Functions

Exported Functions

Index

ModiaMedia.density

ModiaMedia.density_pTX

ModiaMedia.density_ph

ModiaMedia.density_phX

ModiaMedia.dynamicViscosity

ModiaMedia.gasConstant

ModiaMedia.getMedium

ModiaMedia.isenthalpicState

ModiaMedia.isenthalpicState!

ModiaMedia.listMedia

ModiaMedia.pressure

ModiaMedia.pressure_dT

ModiaMedia.saturationPressureOfLiquidWater

ModiaMedia.saturationPressureOfWater

ModiaMedia.setState_dT

ModiaMedia.setState_dT!

ModiaMedia.setState_dTX

ModiaMedia.setState_dTX!

ModiaMedia.setState_pT

ModiaMedia.setState_pT!

ModiaMedia.setState_pTX

ModiaMedia.setState_pTX!

ModiaMedia.setState_ph

ModiaMedia.setState_ph!

ModiaMedia.setState_phX

ModiaMedia.setState_phX!

ModiaMedia.setState_ps

ModiaMedia.setState_ps!

ModiaMedia.setState_psX

ModiaMedia.setState_psX!

ModiaMedia.specificEnthalpy

ModiaMedia.specificEnthalpy_dT

ModiaMedia.specificEnthalpy_pTX

ModiaMedia.specificHeatCapacityCp

ModiaMedia.specificInternalEnergy

ModiaMedia.standardCharacteristics

ModiaMedia.standardPlot

ModiaMedia.sublimationPressureIce

ModiaMedia.temperature

ModiaMedia.temperature_ph

ModiaMedia.temperature_phX

Documentation

ModiaMedia.density — Method.

density(state)

Return density from state::ThermodynamicState in [kg/m^3]

source

ModiaMedia.density_pTX — Method.

density_pTX(medium,p,T,X)

Return density for medium::AbstractMedium from p, T, and X or Xi in [kg/m^3]

source

ModiaMedia.density_ph — Method.

density_ph(medium,p,h)

Return density in [kg/m^3] for medium::PureSubstance from pressure p [Pa] and specific enthalpy h [J/kg].

source

ModiaMedia.density_phX — Method.

density_phX(medium,p,h,X)

Return density for medium::AbstractMedium from p, h, and X or Xi in [kg/m^3]

source

ModiaMedia.dynamicViscosity — Method.

dynamicViscosity(state)

Return dynamic viscosity from state::ThermodynamicState in [Pa*s]

source

ModiaMedia.gasConstant — Method.

R = gasConstant(state)

Return gas constant of MixtureMedium from state::MixtureThermodynamicState in [J/mol.K]

source

ModiaMedia.getMedium — Method.

Medium = getMedium(name::AbstractString)

Return Medium object from medium name. Possible values of argument name can be inquired via listMedia() (Available media).

Examples

medium = getMedium("SimpleAir")

source

ModiaMedia.isenthalpicState! — Method.

isenthalpicState!(state_b,state_a,dp)

Update state_b by an isenthalpic transformation of state_a with pressure drop dp:

pressure(state_b)         = pressure(state_a) + dp
specificEnthalpy(state_b) = specificEnthalpy(state_a)
state_b.X                 = state_a.X

source

ModiaMedia.isenthalpicState — Method.

state_b = isenthalpicState(state_a,dp)

Return state_b by an isenthalpic transformation of state_a with pressure drop dp:

pressure(state_b)         = pressure(state_a) + dp
specificEnthalpy(state_b) = specificEnthalpy(state_a)
state_b.X                 = state_a.X

source

ModiaMedia.listMedia — Method.

listMedia()

List available media of ModiaMedia.

source

ModiaMedia.pressure — Method.

pressure(state)

Return pressure from state::ThermodynamicState in [Pa]

source

ModiaMedia.pressure_dT — Method.

pressure_dT(medium,d,T)

Return pressure in [Pa] for medium::PureSubstance from density d in [kg/m^3] and temperature T [K].

source

ModiaMedia.saturationPressureOfLiquidWater — Method.

saturationPressureOfLiquidWater(Tsat)

Return saturation pressure of liquid water as a function of saturation temperature Tsat in the range of 273.16 to 647.096 K

source

ModiaMedia.saturationPressureOfWater — Method.

saturationPressureOfWater(Tsat)

Return saturation pressure of water as a function of saturation temperature Tsat between 190 K and 647.096 K

source

ModiaMedia.setState_dT! — Method.

setState_dT!(state, d,T)

Update the state::ThermodynamicState of a PureSubstance medium with density d [kg/m^3] and temperature T [K].

source

ModiaMedia.setState_dT — Method.

state = setState_dT(medium, d,T)

Generate a state object for medium medium::PureSubstance for density d [kg/m^3] and temperature T [K].

source

ModiaMedia.setState_dTX! — Method.

setState_dTX!(state, d,T,X)

Update the state::ThermodynamicState object with density d [kg/m^3], temperature T [K] and mass fractions vector X or Xi.

source

ModiaMedia.setState_dTX — Method.

state = setState_dTX(medium, d,T,X)

Generate a state object for medium medium::AbstractMedium for density d [kg/m^3], temperature T [K] and mass fractions vector X or Xi.

source

ModiaMedia.setState_pT! — Method.

setState_pT!(state, p,T)

Update the state::ThermodynamicState of a PureSubstance medium with pressure p [Pa] and temperature T [K].

source

ModiaMedia.setState_pT — Method.

state = setState_pT(medium, p,T)

Generate a state object for medium medium::PureSubstance for pressure p [Pa] and temperature T [K].

source

ModiaMedia.setState_pTX! — Method.

setState_pTX!(state, p,T,X)

Update the state::ThermodynamicState object with pressure p [Pa], temperature T [K] and mass fractions vector X or Xi.

source

ModiaMedia.setState_pTX — Method.

state = setState_pTX(medium, p,T,X)

Generate a state object for medium medium::AbstractMedium for pressure p [Pa], temperature T [K] and mass fractions vector X or Xi.

source

ModiaMedia.setState_ph! — Method.

setState_ph!(state, p,h)

Update the state::ThermodynamicState of a PureSubstance medium with pressure p [Pa] and specific enthalpy h [J/kg]].

source

ModiaMedia.setState_ph — Method.

state = setState_ph(medium, p,h)

Generate a state object for medium medium::PureSubstance for pressure p [Pa] and specific enthalpy h [J/kg].

source

ModiaMedia.setState_phX! — Method.

setState_phX!(state, p,h,X)

Update the state::ThermodynamicState object with pressure p [Pa], specific enthalpy h [J/kg]] and mass fractions vector X or Xi.

source

ModiaMedia.setState_phX — Method.

state = setState_phX(medium, p,h,X)

Generate a state object for medium medium::AbstractMedium for pressure p [Pa], specific enthalpy h [J/kg]] and mass fractions vector X or Xi.

source

ModiaMedia.setState_ps! — Method.

setState_ps!(state, p,s)

Update the state::ThermodynamicState of a PureSubstance medium with pressure p [Pa] and specific entropy s [J/(kg*K)].

source

ModiaMedia.setState_ps — Method.

state = setState_ps(medium, p,s)

Generate a state object for medium medium::PureSubstance for pressure p [Pa] and specific entropy s [J/(kg*K)].

source

ModiaMedia.setState_psX! — Method.

setState_psX!(state, p,s,X)

Update the state::ThermodynamicState object with pressure p [Pa], specific entropy s [J/(kg*K)] and mass fractions vector X or Xi.

source

ModiaMedia.setState_psX — Method.

state = setState_psX(medium, p,s,X)

Generate a state object for medium medium::AbstractMedium for pressure p [Pa], specific entropy s [J/(kg*K)] and mass fractions vector X or Xi.

source

ModiaMedia.specificEnthalpy — Method.

specificEnthalpy(state)

Return specific enthalpy from state::ThermodynamicState in [J/kg]

source

ModiaMedia.specificEnthalpy_dT — Method.

specificEnthalpy_dT(medium,d,T)

Return specific enthalpy in [J/kg] for medium::PureSubstance from density d [kg/m^3] and and temperature T [K].

source

ModiaMedia.specificEnthalpy_pTX — Method.

specificEnthalpy_pTX(medium,p,T,X)

Return specific enthalpy for medium::AbstractMedium from p, T, and X or Xi in [J/kg].

source

ModiaMedia.specificHeatCapacityCp — Method.

specificHeatCapacityCp(state)

Return specific heat capacity at constant pressure from state::ThermodynamicState in [J/(kg*K)]

source

ModiaMedia.specificInternalEnergy — Method.

specificInternalEnergy(state)

Return specific internal energy from state::ThermodynamicState in [J/kg]

source

ModiaMedia.standardCharacteristics — Method.

dict = standardCharacteristics(medium::AbstractMedium)

Return a dict::dict{AbstractString,Any} dictionary with the most important characteristics of the medium as vectors.

source

ModiaMedia.standardPlot — Method.

standardPlot(medium::AbstractMedium; figure=1)

Plot the standardCharacteristics(medium) of the medium.

source

ModiaMedia.sublimationPressureIce — Method.

sublimationPressureIce(Tsat)

Return sublimation pressure of water as a function of saturation temperature Tsat between 190 and 273.16 K

source

ModiaMedia.temperature — Method.

temperature(state)

Return temperature from state::ThermodynamicState in [K]

source

ModiaMedia.temperature_ph — Method.

temperature_ph(medium,p,h)

Return temperature in [K] for medium::PureSubstance from pressure p [Pa] and specific enthalpy h [J/kg].

source

ModiaMedia.temperature_phX — Method.

temperature_phX(medium,p,h,X)

Return temperature for medium::AbstractMedium from p, h, and X or Xi in [K].

source