Functions

Functions

ModiaMedia.cp_T — Method.

cp = cp_T(data::SingleGasNasaData, T) - Compute specific heat capacity at constant pressure from temperature and gas data

source

ModiaMedia.density — Method.

d = density(medium,state) - return density for medium from state in [kg/m^3]

source

ModiaMedia.density_pTX — Method.

d = density_pTX(medium,p,T,X) - return density for medium from p, T, and X or Xi in [kg/m^3]

source

ModiaMedia.density_ph — Method.

d = density_ph(medium,p,h) - return density in [kg/m^3] for medium::PureSubstance from pressure p [Pa] and specific enthalpy h [J/kg].

source

ModiaMedia.density_phX — Method.

d = density_phX(medium,p,h,X) - return density for medium from p, h, and X or Xi in [kg/m^3]

source

ModiaMedia.h_T — Function.

h = h_T(data, T, exclEnthForm=true, refChoice=ReferenceEnthalpy_ZeroAt0K, h_off=0.0)

Return specific enthalpy from temperature and gas data.

Arguments

data::SingleGasNasaData: Data of the SingleGasNasa medium.

T::Float64: Temperature in [K].

exclEnthForm::Bool: If true, enthalpy of formation Hf is not included in specific enthalpy h.

refChoice::ReferenceEnthalpy: Choice of reference enthalpy.

h_off::Float64: User defined offset for reference enthalpy, if referenceEnthalpy = ReferenceEnthalpy_UserDefined

source

ModiaMedia.pressure — Method.

p = pressure(medium,state) - return pressure for medium from state in [Pa]

source

ModiaMedia.pressure_dT — Method.

p = pressure_dT(medium,d,T) - return pressure in [Pa] for medium::PureSubstance from density c in [kg/m^3] and temperature T [K].

source

ModiaMedia.s0_T — Method.

Compute specific entropy from temperature and gas data

source

ModiaMedia.setState_dT — Method.

state = setState_dT(medium, d,T)

Generate a state object for medium medium::PureSubstance for density d [kg/m^3] and temperature T [K].

source

ModiaMedia.setState_dTX — Method.

state = setState_dTX(medium, d,T,X)

Generate a state object for medium medium::AbstractMedium for density d [kg/m^3], temperature T [K] and mass fractions vector X or Xi.

source

ModiaMedia.setState_pT — Method.

state = setState_pT(medium, p,T)

Generate a state object for medium medium::PureSubstance for pressure p [Pa] and temperature T [K].

source

ModiaMedia.setState_pTX — Method.

state = setState_pTX(medium, p,T,X)

Generate a state object for medium medium::AbstractMedium for pressure p [Pa], temperature T [K] and mass fractions vector X or Xi.

source

ModiaMedia.setState_ph — Method.

state = setState_ph(medium, p,h)

Generate a state object for medium medium::PureSubstance for pressure p [Pa] and specific enthalpy h [J/kg].

source

ModiaMedia.setState_phX — Method.

state = setState_phX(medium, p,h,X)

Generate a state object for medium medium::AbstractMedium for pressure p [Pa], specific enthalpy h [J/kg]] and mass fractions vector X or Xi.

source

ModiaMedia.setState_ps — Method.

state = setState_ps(medium, p,s)

Generate a state object for medium medium::PureSubstance for pressure p [Pa] and specific entropy s [J/(kg*K)].

source

ModiaMedia.setState_psX — Method.

state = setState_psX(medium, p,s,X)

Generate a state object for medium medium::AbstractMedium for pressure p [Pa], specific entropy s [J/(kg*K)] and mass fractions vector X or Xi.

source

ModiaMedia.specificEnthalpy — Method.

h = specificEnthalpy(medium,state) - return specific enthalpy for medium from state in [J/kg]

source

ModiaMedia.specificEnthalpy_dT — Method.

h = specificEnthalpy_dT(medium,d,T) - return specific enthalpy in [J/kg] for medium from density d [kg/m^3] and and temperature T [K].

source

ModiaMedia.specificEnthalpy_pTX — Method.

h = specificEnthalpy_pTX(medium,p,T,X) - return specific enthalpy for medium from p, T, and X or Xi in [J/kg]

source

ModiaMedia.specificHeatCapacityCp — Method.

cp = specificHeatCapacityCp(medium,state) - return specific heat capacity at constant pressure for medium from state in [J/(kg*K)]

source

ModiaMedia.specificInternalEnergy — Method.

u = specificInternalEnergy(medium,state) - return specific internal energy from medium at state in [J/kg]

source

ModiaMedia.standardCharacteristics — Method.

dict = standardCharacteristics(medium::AbstractMedium)

Return a dict::dict{AbstractString,Any} dictionary with the most important characteristics of the medium as vectors.

source

ModiaMedia.standardPlot — Method.

standardPlot(medium::AbstractMedium; figure=1)

Plot the standardCharacteristics(medium) of the medium.

source

ModiaMedia.temperature — Method.

T = temperature(medium,state) - return temperature for medium from state in [K]

source

ModiaMedia.temperature_ph — Method.

T = temperature_ph(medium,p,h) - return temperature in [K] for medium::PureSubstance from pressure p [Pa] and specific enthalpy h [J/kg].

source

ModiaMedia.temperature_phX — Method.

T = temperature_phX(medium,p,h,X) - return temperature for medium from p, h, and X or Xi in [K]

source